% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bean_methods.R
\docType{class}
\name{bean_optimization}
\alias{bean_optimization}
\title{bean_optimization object}
\description{
An object of class \code{bean_optimization} is a list returned by the
\code{\link{find_optimal_cap}} function. It contains the results of the
optimization search.
}
\keyword{internal}
