% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{occ_data_raw}
\alias{occ_data_raw}
\title{Raw Rusa unicolor occurrence data}
\format{
A data frame with occurrence coordinates:
\describe{
  \item{species}{The species name}
  \item{x}{Longitude coordinates in decimal degrees}
  \item{y}{Latitude coordinates in decimal degrees}
}
}
\source{
Example dataset provided with the bean package.
}
\usage{
occ_data_raw
}
\description{
A dataset containing the raw, unthinned occurrence records for Rusa unicolor
(Sambar deer) in Thailand. This dataset is used to demonstrate the spatial
clustering and environmental bias present in typical SDM datasets.
}
\keyword{datasets}
