% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{origin_ellipse}
\alias{origin_ellipse}
\title{Fitted niche ellipsoid for Rusa unicolor}
\format{
A \code{bean_ellipsoid} object (see \code{\link{fit_ellipsoid}}).
}
\usage{
origin_ellipse
}
\description{
A \code{bean_ellipsoid} fitted to \code{\link{origin_dat_prepared}}
representing the baseline environmental niche of the species.
}
\keyword{datasets}
