% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_maniculatus_sample}
\alias{P_maniculatus_sample}
\title{Occurrence data for the deer mouse, Peromyscus maniculatus.}
\format{
A data frame with rows for each occurrence and columns for:
}
\source{
\url{https://www.gbif.org/}
}
\usage{
P_maniculatus_sample
}
\description{
A dataset containing species occurrences and associated environmental variables
from the GBIF database.
}
\keyword{datasets}
