% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_env_thinning.R
\name{plot.bean_evaluation}
\alias{plot.bean_evaluation}
\title{Plot bean_evaluation results}
\usage{
\method{plot}{bean_evaluation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bean_evaluation}.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object.
}
\description{
Creates a diagnostic plot from the output of \code{\link{test_env_thinning}}.
}
\keyword{internal}
