% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_cap.R
\name{plot.bean_optimization}
\alias{plot.bean_optimization}
\title{Plot bean_optimization results}
\usage{
\method{plot}{bean_optimization}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bean_optimization}.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object.
}
\description{
Creates a diagnostic plot from the output of \code{\link{find_optimal_cap}}.
}
