% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_env_resolution.R
\name{plot.bean_resolution}
\alias{plot.bean_resolution}
\title{Plot bean_resolution results}
\usage{
\method{plot}{bean_resolution}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bean_resolution}.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object.
}
\description{
Creates a diagnostic plot from the output of \code{\link{find_env_resolution}}.
}
