% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_env_nd.R
\name{print.bean_thinned}
\alias{print.bean_thinned}
\title{Print a summary of bean_thinned results}
\usage{
\method{print}{bean_thinned}(x, ...)
}
\arguments{
\item{x}{An object of class `bean_thinned`.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object `x`.
}
\description{
Print a summary of bean_thinned results
}
\keyword{internal}
