% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_env_density.R
\name{thin_env_density}
\alias{thin_env_density}
\title{Thin occurrence data based on environmental density (stochastic)}
\usage{
thin_env_density(data, env_vars, grid_resolution, max_per_cell, verbose = TRUE)
}
\arguments{
\item{data}{A data frame containing species occurrences and environmental data.}

\item{env_vars}{A character vector of length two specifying the names of the
environmental variables to use.}

\item{grid_resolution}{A numeric vector of length one or two specifying the
resolution(s) for the grid axes. If length one, it is used for both axes.}

\item{max_per_cell}{An integer specifying the maximum number of points to
retain per grid cell.}

\item{verbose}{(logical) If TRUE, prints progress messages. Default = TRUE.}
}
\value{
An object of class \code{bean_thinned_density}.
}
\description{
This function thins species occurrence records in a 2D environmental
space by randomly sampling a specified number of points from each occupied
grid cell.
}
