% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{thinned_deterministic}
\alias{thinned_deterministic}
\title{Deterministically thinned environmental data}
\format{
A \code{bean_thinned_center} object (see
  \code{\link{thin_env_center}}).
}
\usage{
thinned_deterministic
}
\description{
Result of \code{\link{thin_env_center}} applied to
\code{\link{origin_dat_prepared}}. Contains one calculated centroid per
occupied environmental grid cell.
}
\keyword{datasets}
