% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{thinned_stochastic}
\alias{thinned_stochastic}
\title{Stochastically thinned environmental data}
\format{
A \code{bean_thinned} object (see \code{\link{thin_env_nd}}).
}
\usage{
thinned_stochastic
}
\description{
Result of \code{\link{thin_env_nd}} applied to
\code{\link{origin_dat_prepared}}. Contains one randomly chosen occurrence
per occupied environmental grid cell.
}
\keyword{datasets}
