% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bean-package.R
\docType{package}
\name{bean-package}
\alias{bean}
\alias{bean-package}
\title{bean: Data Thinning of Species Occurrences in Environmental Space}
\description{
Provides tools to address sampling bias in species occurrence records by thinning data in environmental space (E-space), a crucial pre-processing step for ecological niche modeling (ENM). The package offers a data-driven protocol to objectively determine thinning parameters, provides both stochastic and deterministic thinning methods, and includes functions to delineate the resulting environmental niche using robust ellipses. The name 'bean' reflects the core principle of the method: ensuring that each “pod” (a grid cell in E-space) contains only a specified number of “beans” (occurrence points).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paanwaris/bean}
  \item Report bugs at \url{https://github.com/paanwaris/bean/issues}
}

}
\author{
\strong{Maintainer}: Paanwaris Paansri \email{paanwaris@vt.edu}

Authors:
\itemize{
  \item Luis E. Escobar \email{escobar1@vt.edu}
}

}
\keyword{internal}
