% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bean-package.R
\docType{package}
\name{bean-package}
\alias{bean}
\alias{bean-package}
\title{bean: Data Thinning of Species Occurrences in Environmental Space}
\description{
Provides a suite of tools to mitigate sampling bias in species occurrence records by thinning data in the environmental space (E-space). This process could help increase accuracy and precision in species distribution modeling (SDM) and ecological niche modeling (ENM). The package offers a data-driven protocol to determine thinning parameters. Thinning methods (stochastic and deterministic) help reduce oversampled conditions and outlier observations. The name 'bean' reflects the core principle of the method: ensuring that each “pod” (a grid cell in E-space) contains only a specified number of “beans” (occurrence points).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paanwaris/bean}
  \item Report bugs at \url{https://github.com/paanwaris/bean/issues}
}

}
\author{
\strong{Maintainer}: Paanwaris Paansri \email{paanwaris@vt.edu}

Other contributors:
\itemize{
  \item Luis E. Escobar \email{escobar1@vt.edu} [contributor]
}

}
\keyword{internal}
