% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bean-package.R
\docType{package}
\name{bean-package}
\alias{bean}
\alias{bean-package}
\title{bean: Spatially-Aware Environmental Thinning of Species Occurrence Data}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paanwaris/bean}
  \item Report bugs at \url{https://github.com/paanwaris/bean/issues}
}

}
\author{
\strong{Maintainer}: Paanwaris \email{paanwaris_04@hotmail.com}

}
\keyword{internal}
