% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_env_resolution.R
\name{find_env_resolution}
\alias{find_env_resolution}
\title{Find an objective grid resolution using pairwise environmental distances}
\usage{
find_env_resolution(data, env_vars, quantile = 0.1, verbose = TRUE)
}
\arguments{
\item{data}{A data frame containing species occurrences and pre-scaled
environmental data.}

\item{env_vars}{A character vector of length two specifying the names of the
environmental variables to use.}

\item{quantile}{(numeric) The quantile of pairwise distances to use for the
resolution. A smaller value (e.g., 0.05 or 0.1) is recommended.
Default = 0.1.}

\item{verbose}{(logical) If TRUE, prints progress messages and warnings.
Default = TRUE.}
}
\value{
An object of class \code{bean_resolution}.
}
\description{
This function assumes the environmental variables in the input
data have already been scaled (e.g., using `scale()`). It calculates the
distribution of all pairwise Euclidean distances between points in this
environmental space. The distance at a specified lower quantile of this
distribution is returned as a suggested grid resolution.
}
