% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_cap.R
\name{find_optimal_cap}
\alias{find_optimal_cap}
\title{Find optimal density caps based on a target thinning percentage}
\usage{
find_optimal_cap(
  data,
  env_vars,
  grid_resolution,
  target_percent,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing species occurrences and environmental data.}

\item{env_vars}{A character vector of length two specifying the names of the
columns to be used as the axes of the environmental space.}

\item{grid_resolution}{A numeric vector of length one or two specifying the
resolution(s) for the grid axes. If length one, it is used for both axes.}

\item{target_percent}{A numeric value (0-1) for the target proportion of
points to retain.}

\item{verbose}{(logical) If TRUE, prints progress messages. Default = TRUE.}
}
\value{
An object of class \code{bean_optimization}.
}
\description{
This function searches for optimal density caps using two criteria:
1) the cap that results in a point count closest to the target percentage, and
2) the cap that results in the closest point count AT OR ABOVE the target.
}
