% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_cap.R
\name{find_optimal_cap}
\alias{find_optimal_cap}
\title{Find optimal density caps based on a target thinning percentage}
\usage{
find_optimal_cap(data, env_vars, grid_resolution, target_percent, seed = NULL)
}
\arguments{
\item{data}{A data frame containing species occurrences and environmental data.}

\item{env_vars}{A character vector of length two specifying the names of the
columns to be used as the axes of the environmental space.}

\item{grid_resolution}{A single numeric value specifying the resolution of the
grid.}

\item{target_percent}{A numeric value (0-1) for the target proportion of
points to retain.}

\item{seed}{An optional integer to set the random seed for reproducibility.}
}
\value{

}
\description{
This function searches for optimal density caps using two criteria:
1) the cap that results in a point count closest to the target percentage, and
2) the smallest cap that retains at least the target percentage.
}
