% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ellipsoid.R
\name{fit_ellipsoid}
\alias{fit_ellipsoid}
\title{Fit a bivariate environmental ellipsoid to occurrence data}
\usage{
fit_ellipsoid(data, var1, var2, method = "covmat", level = 95)
}
\arguments{
\item{data}{A data frame containing scaled environmental values and coordinates.}

\item{var1}{The name of the first environmental variable (e.g., "BIO1").}

\item{var2}{The name of the second environmental variable (e.g., "BIO12").}

\item{method}{Method used to estimate the ellipse: "covmat", "mve1", or "mve2".}

\item{level}{Confidence level (e.g., 95).}
}
\value{
An object of class \code{bean_ellipsoid}.
}
\description{
Fit a bivariate environmental ellipsoid to occurrence data
}
