% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{origin_ellipse}
\alias{origin_ellipse}
\title{Fitted origin ellipsoid}
\format{
An object of class \code{bean_ellipsoid} containing:
\describe{
  \item{centroid}{The multidimensional mean of the environmental variables}
  \item{covariance_matrix}{The covariance matrix of the variables}
  \item{volume}{The hypervolume of the fitted ellipsoid}
  \item{parameters}{Model parameters including the confidence level (e.g., 0.95)}
}
}
\usage{
origin_ellipse
}
\description{
A multidimensional ellipsoid fitted to the unthinned, prepared occurrence data
representing the baseline environmental niche of the species.
}
\keyword{internal}
