% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ellipsoid.R
\name{plot.bean_ellipsoid}
\alias{plot.bean_ellipsoid}
\title{Plot bean_ellipsoid results}
\usage{
\method{plot}{bean_ellipsoid}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bean_ellipsoid}.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object.
}
\description{
Creates a diagnostic plot from the output of \code{\link{fit_ellipsoid}}.
}
