% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_env_center.R
\name{print.bean_thinned_center}
\alias{print.bean_thinned_center}
\title{Print a summary of bean_thinned_center results}
\usage{
\method{print}{bean_thinned_center}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bean_thinned_center}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print a summary of bean_thinned_center results
}
\keyword{internal}
