% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_env_density.R
\name{print.bean_thinned_density}
\alias{print.bean_thinned_density}
\title{Print a summary of bean_thinned_density results}
\usage{
\method{print}{bean_thinned_density}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bean_thinned_density}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print a summary of bean_thinned_density results
}
\keyword{internal}
