% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_env_center.R
\name{thin_env_center}
\alias{thin_env_center}
\title{Thin occurrence data to grid cell centers (deterministic)}
\usage{
thin_env_center(data, env_vars, grid_resolution, verbose = TRUE)
}
\arguments{
\item{data}{A data frame containing species occurrences and environmental data.}

\item{env_vars}{A character vector of length two specifying the names of the
environmental variables to use.}

\item{grid_resolution}{A numeric vector of length one or two specifying the
resolution(s) for the grid axes. If length one, it is used for both axes.}

\item{verbose}{(logical) If TRUE, prints progress messages. Default = TRUE.}
}
\value{
An object of class \code{bean_thinned_center}.
}
\description{
This function thins species occurrence records by finding all
occupied cells in a 2D environmental grid and returning a single new point
at the exact center of each of those cells. This is a deterministic method.
}
