% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_env_density.R
\name{thin_env_density}
\alias{thin_env_density}
\title{Stochastic centroid}
\usage{
thin_env_density(data, env_vars, grid_resolution, max_per_cell)
}
\arguments{
\item{data}{A data.frame containing species occurrence coordinates and the environmental variables.}

\item{env_vars}{A character vector specifying the column names in data that represent the environmental variables to be used in the analysis.}

\item{grid_resolution}{A numeric vector of length one or two specifying the
resolution(s) for the grid axes. If length one, it is used for both axes. See the Details section of \code{\link{find_env_resolution}} for a full explanation}

\item{max_per_cell}{An integer specifying the maximum number of points to
retain per grid cell. See the Details section of \code{\link{find_optimal_cap}} for a full explanation}
}
\value{
An object of class \code{bean_thinned_density}, which is a list containing:
  \item{thinned_data}{A data.frame containing the occurrence records that were retained after the thinning process.}
  \item{n_original}{An integer representing the number of complete occurrence records in the input data before thinning.}
  \item{n_thinned}{An integer representing the number of occurrence records remaining after thinning.}
}
\description{
This function thins species occurrence records in a 2D environmental
space by randomly sampling a specified number of points from each occupied
grid cell. This is a stochastic method.
}
\examples{
\dontrun{
# 1. Create environmental data
env_data <- data.frame(
BIO1 = c(rep(0.1, 5), rep(1.1, 3)),
BIO12 = c(rep(0.2, 5), rep(2.2, 3)),
species = "A"
)

# 2. Thin the data to a max of 1 point per cell
set.seed(81) # For reproducible results
thinned_obj <- thin_env_density(
  data = env_data,
  env_vars = c("BIO1", "BIO12"),
  grid_resolution = c(0.2, 0.2),
  max_per_cell = 1
)

# 3. Print the summary
print(thinned_obj)
}
}
\seealso{
\code{\link{find_optimal_cap}}, \code{\link{find_env_resolution}}
}
