% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{thinned_deterministic}
\alias{thinned_deterministic}
\title{Deterministic thinned environmental data}
\format{
An object of class \code{bean_thinned_center} containing:
\describe{
  \item{thinned_points}{A data frame of the new geometric center points}
  \item{n_original}{The number of points before thinning}
  \item{n_thinned}{The number of unique grid cells occupied}
  \item{parameters}{The thinning parameters used}
}
}
\usage{
thinned_deterministic
}
\description{
The output object generated by running \code{\link{thin_env_center}} on the
prepared occurrence data. It contains a single calculated centroid coordinate
for each occupied environmental grid cell.
}
\keyword{datasets}
