% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{thinned_stochastic}
\alias{thinned_stochastic}
\title{Stochastic thinned environmental data}
\format{
An object of class \code{bean_thinned_density} containing:
\describe{
  \item{thinned_data}{A data frame of the retained occurrence points}
  \item{n_original}{The number of points before thinning}
  \item{n_thinned}{The number of points after thinning}
  \item{parameters}{The thinning parameters used}
}
}
\usage{
thinned_stochastic
}
\description{
The output object generated by running \code{\link{thin_env_nd}} on the
prepared occurrence data. It contains exactly one randomly selected point
per occupied environmental grid cell.
}
\keyword{datasets}
